modify playlist invite code

+++ src/app/api/user/signOut/route.ts
@@ -7,12 +7,10 @@ export async function POST(req: NextRequest) {
 
   const authSupabase = makeServerClient();
   const { error: authError } = await authSupabase.auth.signOut();
+
   if (authError) return NextResponse.json("no user", { status: 401 });
 
-  const { data, error } = await supabase
-    .from("token")
-    .delete()
-    .eq("user_id", userId as string);
+  const { data, error } = await supabase.from("token").delete().eq("user_id", userId);
 
   if (error) return NextResponse.json({ error: error.message }, { status: 500 });
   return NextResponse.json(data);
